Refactoring of menu content
