+++ assets/js/components/tracks_manager.ts
@@ -18,6 +18,7 @@ import {
   getContainer,
   getURLRow,
 } from "./util/menu_utils";
+import { getAnnotationContextMenuContent, getGenesContextMenuContent, getVariantContextMenuContent } from "./util/menu_content_utils";
 
 const COV_Y_RANGE: [number, number] = [-4, 4];
 const BAF_Y_RANGE: [number, number] = [0, 1];
@@ -109,7 +110,7 @@ export class TracksManager extends HTMLElement {
       (range) => {
         this.highlights.push(range);
         this.render(false);
-      }
+      },
     );
     const bafTrack = new DotTrack(
       "baf",
@@ -129,7 +130,7 @@ export class TracksManager extends HTMLElement {
       (range) => {
         this.highlights.push(range);
         this.render(false);
-      }
+      },
     );
     const variantTrack = new BandTrack(
       "variants",
@@ -153,7 +154,7 @@ export class TracksManager extends HTMLElement {
       (range) => {
         this.highlights.push(range);
         this.render(false);
-      }
+      },
     );
     const genesTrack = new BandTrack(
       "genes",
@@ -176,7 +177,7 @@ export class TracksManager extends HTMLElement {
       (range) => {
         this.highlights.push(range);
         this.render(false);
-      }
+      },
     );
     const ideogramTrack = new IdeogramTrack(
       "ideogram",
@@ -211,7 +212,6 @@ export class TracksManager extends HTMLElement {
         track.style.paddingLeft = `${STYLE.yAxis.width}px`;
         return track;
       },
-
     );
 
     const overviewTrackCov = new OverviewTrack(
@@ -285,6 +285,18 @@ function getAnnotTrack(
   setXRange: (range: Rng) => void,
   onZoomOut: () => void,
 ): BandTrack {
+  async function getAnnotTrackData(
+    source: string,
+    getXRange: () => Rng,
+    getAnnotation: (source: string) => Promise<RenderBand[]>,
+  ): Promise<BandTrackData> {
+    const bands = await getAnnotation(source);
+    return {
+      xRange: getXRange(),
+      bands,
+    };
+  }
+
   const track = new BandTrack(
     sourceId,
     label,
@@ -297,125 +309,9 @@ function getAnnotTrack(
     (range) => {
       this.highlights.push(range);
       this.render(false);
-    }
+    },
   );
   return track;
-}
-
-async function getAnnotTrackData (
-  source: string,
-  getXRange: () => Rng,
-  getAnnotation: (source: string) => Promise<RenderBand[]>,
-): Promise<BandTrackData> {
-  const bands = await getAnnotation(source);
-  return {
-    xRange: getXRange(),
-    bands,
-  };
-};
-
-function getVariantContextMenuContent(id: string, details: ApiVariantDetails, variantUrl: string): HTMLDivElement[] {
-  const info: { key: string; value: string; url?: string }[] = [
-    { key: "Range", value: `${details.position} - ${details.end}` },
-    {
-      key: "Length",
-      value: prefixNts(details.length),
-    },
-    { key: "Variant URL", value: variantUrl, url: variantUrl },
-    { key: "CADD score", value: details.cadd_score },
-    {
-      key: "Category",
-      value: `${details.category} (${details.sub_category})`,
-    },
-    {
-      key: "Cytoband",
-      value: `${details.cytoband_start} - ${details.cytoband_end}`,
-    },
-    { key: "Rank score", value: details.rank_score.toString() },
-    { key: "BNF ID", value: id },
-  ];
-
-  const entries: HTMLDivElement[] = info.map((i) => getEntry(i));
-  const rankScoreParts = getSection(
-    "Rank score parts",
-    details.rank_score_results.map((part) => {
-      return getContainer("row", `${part.category}: ${part.score}`);
-    }),
-  );
-
-  entries.push(rankScoreParts);
-  return entries;
-}
-
-function getGenesContextMenuContent(id: string, details: ApiTranscriptDetails): HTMLDivElement[] {
-  const info: { key: string; value: string }[] = [
-    { key: "Range", value: `${details.start} - ${details.end}` },
-    {
-      key: "Length",
-      value: prefixNts(details.end - details.start + 1),
-    },
-    { key: "Transcript ID", value: details.transcript_id },
-    { key: "Biotype", value: details.transcript_biotype },
-    { key: "Gene name", value: details.gene_name },
-    { key: "MANE", value: details.mane },
-    { key: "HGNC ID", value: details.hgnc_id },
-    { key: "Refseq ID", value: details.refseq_id },
-    { key: "Strand", value: details.strand },
-    { key: "BNF ID", value: id },
-  ];
-
-  const entries = info.map((i) => getEntry(i));
-  return entries;
-}
-
-function getAnnotationContextMenuContent(id: string, details: ApiAnnotationDetails): HTMLDivElement[] {
-  const info: { key: string; value: string }[] = [
-    {
-      key: "Range",
-      value: prettyRange(details.start, details.end),
-    },
-    {
-      key: "Length",
-      value: prefixNts(details.end - details.start + 1),
-    },
-    { key: "Name", value: details.name },
-    { key: "Description", value: details.description || "-" },
-    { key: "BNF ID", value: id },
-  ];
-
-  const infoDivs = info.map((i) => getEntry(i));
-
-  const commentSection = getSection(
-    "Comments",
-    details.comments.flatMap((c) =>
-      c.comment
-        .replace('"', "")
-        .split("; ")
-        .map((s) => getURLRow(s)),
-    ),
-  );
-  infoDivs.push(commentSection);
-
-  const metaSection = getSection(
-    "Metadata",
-    details.metadata.map((meta) => {
-      let url = null;
-      if (meta.field_name === "reference") {
-        const value = meta.value as {url: string, title: string};
-        url = value.url;
-      }
-      
-      const entry = getEntry({
-        key: meta.field_name,
-        url,
-        value: url != null ? url : meta.value as string
-      })
-      return entry;
-    }
-    ),
-  );
-  infoDivs.push(metaSection);
-  return infoDivs;
 }
 
 customElements.define("gens-tracks", TracksManager);
